Remove SiteFooter component and move meaningful parts into FooterBlock
