+++ apps/store/src/blocks/FooterBlock.tsx
@@ -1,3 +1,4 @@
+import styled from '@emotion/styled'
 import { storyblokEditable } from '@storyblok/react'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
@@ -5,17 +6,8 @@ import { ChangeEvent, useRef } from 'react'
 import { Space } from 'ui'
 import * as Accordion from '@/components/Accordion/Accordion'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
-import {
-  Flex,
-  SiteFooterProps,
-  StyledAccordionContent,
-  StyledLink,
-  TextMuted,
-  Wrapper,
-} from '@/components/SiteFooter/SiteFooter'
-import { Field, MARKET_MAP, TEMP_TRANSLATIONS } from '@/components/SiteFooter/SiteFooter.constants'
-import { findLocale } from '@/components/SiteFooter/SiteFooter.helpers'
 import { SpaceFlex } from '@/components/SpaceFlex/SpaceFlex'
+import { LocaleField, MARKET_MAP, TEMP_TRANSLATIONS, findLocale } from '@/lib/l10n/locales'
 import { Locale } from '@/lib/l10n/types'
 import { useCurrentLocale } from '@/lib/l10n/useCurrentLocale'
 import { ExpectedBlockType, LinkField, SbBaseBlockProps } from '@/services/storyblok/storyblok'
@@ -59,9 +51,7 @@ FooterSection.blockName = 'footerSection' as const
 
 type FooterBlockProps = SbBaseBlockProps<{
   footerSections: ExpectedBlockType<FooterSectionProps>
-}> &
-  SiteFooterProps
-
+}>
 export const FooterBlock = ({ blok }: FooterBlockProps) => {
   const formRef = useRef<HTMLFormElement>(null)
   const { marketLabel: currentMarket, htmlLang: currentLanguage } = useCurrentLocale()
@@ -86,7 +76,9 @@ export const FooterBlock = ({ blok }: FooterBlockProps) => {
   }
   const handleSubmit = (event: ChangeEvent<HTMLFormElement>) => {
     event.preventDefault()
-    onChangeLocale(findLocale(event.target[Field.Market].value, event.target[Field.Language].value))
+    onChangeLocale(
+      findLocale(event.target[LocaleField.Market].value, event.target[LocaleField.Language].value),
+    )
   }
 
   const footerSections = filterByBlockType(blok.footerSections, FooterSection.blockName)
@@ -102,7 +94,7 @@ export const FooterBlock = ({ blok }: FooterBlockProps) => {
         <SpaceFlex>
           <Flex>
             <InputSelect
-              name={Field.Market}
+              name={LocaleField.Market}
               onChange={handleSelectChange}
               value={currentMarket}
               options={marketList}
@@ -111,7 +103,7 @@ export const FooterBlock = ({ blok }: FooterBlockProps) => {
 
           <Flex>
             <InputSelect
-              name={Field.Language}
+              name={LocaleField.Language}
               onChange={handleSelectChange}
               value={currentLanguage}
               options={languageList}
@@ -125,3 +117,25 @@ export const FooterBlock = ({ blok }: FooterBlockProps) => {
   )
 }
 FooterBlock.blockName = 'footer' as const
+
+export const Wrapper = styled(Space)(({ theme }) => ({
+  width: '100%',
+  backgroundColor: theme.colors.gray200,
+  padding: `${theme.space[6]} ${theme.space[4]}`,
+}))
+
+export const Flex = styled.div({ flex: 1 })
+
+export const TextMuted = styled.p(({ theme }) => ({
+  color: theme.colors.gray600,
+  fontSize: theme.fontSizes[1],
+}))
+
+export const StyledAccordionContent = styled(Accordion.Content)(({ theme }) => ({
+  padding: theme.space[4],
+  paddingTop: theme.space[2],
+}))
+
+export const StyledLink = styled.a({
+  textDecoration: 'none',
+})
