+++ apps/store/src/components/LayoutWithMenu/LayoutWithMenu.tsx
@@ -1,10 +1,8 @@
 import styled from '@emotion/styled'
 import { StoryData } from '@storyblok/react'
-import { useRouter } from 'next/router'
 import { ReactElement } from 'react'
 import { FooterBlock } from '@/blocks/FooterBlock'
 import { HeaderBlock } from '@/blocks/TopMenuBlock'
-import { SiteFooterProps } from '@/components/SiteFooter/SiteFooter'
 
 const Wrapper = styled.div({
   minHeight: '100vh',
@@ -18,11 +16,6 @@ type LayoutWithMenuProps = {
 }
 
 export const LayoutWithMenu = ({ children }: LayoutWithMenuProps) => {
-  const router = useRouter()
-  const handleChangeLocale: SiteFooterProps['onChangeLocale'] = (locale) => {
-    router.push(router.asPath, undefined, { locale })
-  }
-
   const story = children.props.story as StoryData | undefined
   const globalStory = children.props.globalStory
 
@@ -30,9 +23,7 @@ export const LayoutWithMenu = ({ children }: LayoutWithMenuProps) => {
     <Wrapper>
       {(!story || !story.content.hideMenu) && <HeaderBlock blok={globalStory.content} />}
       {children}
-      {(!story || !story.content.hideMenu) && (
-        <FooterBlock blok={globalStory.content} onChangeLocale={handleChangeLocale} />
-      )}
+      {(!story || !story.content.hideMenu) && <FooterBlock blok={globalStory.content} />}
     </Wrapper>
   )
 }
