--- apps/store/src/lib/l10n/getLocale.ts
@@ -1,6 +1,0 @@
-import { FALLBACK_LOCALE, locales } from './locales'
-import { Locale } from './types'
-
-export const getLocale = (locale: Locale | string | undefined) => {
-  return locales[locale as Locale] ?? locales[FALLBACK_LOCALE]
-}
